Use hash-based meta/ on-disk storage
Instead of meta/foo/.DS-Store, the JSON metadata for a file will be stored in
meta/AB/CDEFG…, where sha3(/foo) = ABCDEFG…

Performance-wise, since knowing whether we had to add .DS-Store to the path
meant performing a stat() on the web/ file, the number of files we access is
roughly the same, even when considering file system-level caching. In fact,
performance should improve very slightly for two reasons:

1. The server no longer has to wait for the result of the web/ stat() before it
   can provide the metafile's path, which means it only waits for the metafile's
   content to be loaded.
2. Metadata access only needs to follow a 2-node path (the AB directory and the
   CDEFG… file), instead of a path through the file system that may have an
   arbitrary depth.

However, there is a cons: knowing whether a metafile no longer corresponds to
any file (and can be deleted) requires matching all files to their metafiles and
seeing which metafiles were not matched. Before this change, they could be
looked up directly.