Prevent allowing a script to load private information.
Imagine Alice crafted a Web page on thefiletree.com with a script that
loaded a page from /bob/secret which is ACL'ed to only be readable by
Bob. It is set up so that the credentials are sent with the XHR call to
read that page, and to store the contents to a page that Alice can read.
She maliciously sends the link to Bob, who clicks on it, unknowingly
leaking the contents of /bob/secret to Alice.

Here, we prevent this by blocking Ajax for ?app=data page loads. As a
result, all data that JS code needs must be in the page.

(Ideally we would want to prevent sending credentials instead.)

The idea going forward is to apply this policy for ?app=data, and check
the user's allowed apps for other apps. (Currently, we only allow apps
from /app, which we control.)