Modifications in TODO.md
Hello, I have modified some of the wording in TODO.md. I have refrained from changing instructions that might alter the message. I have one question though. In the line 56, I'm not quite sure what the sentence "We get a salt from the server." means. So, I left it as is but have a feeling it probably needs to rephrased. Thanks!