Add appAuth to template parameters
Imagine that Malicious Mary wants to read a file that Gentle Jane keeps
secret. She therefore crafts an HTML page on thefiletree.com, of which
link she sends to Jane. When Jane opens the page, her browser executes
the JS code, which runs an XHR in the background to fetch the secret
page. That sends over her httpOnly token cookie to thefiletree.com,
which authenticates her. The page cannot see the contents of the
cookies, but it still downloads the secret page, and sends it to Mary.

To avoid that, we need two pieces of information: user identification,
and proof of the source of the call. Indeed, apps that Jane approves
should still be able to perform calls to thefiletree.com on her behalf.
All we'd need is the path of the HTML file making the call; but we
cannot just send the path in a header, as anyone could craft a request
for the path of an app they know Jane approved (eg. a default app, like
`/app/folder`).

However, we can yield a keyed hash (with SHA3 / BLAKE2, or using HMAC)
of the path, keyed by Jane's secret, to approved apps. It does not leak
Jane's secret, and cannot be obtained without knowing Jane's secret.
Also, it is unique to each path, so that one cannot use a well-known
approved path for a call from a malicious page.

    APP_AUTH: /app/folder blake2b256 65AczoAPRpCxq_EIBv8zRIQ8mAnF6qxO8pGhZyUDUdKY

The server can compute the keyed hash from the app path and the secret
received from the cookie, to verify the source of the call. Then, it can
verify that the app is approved for the user.